+++ Buildings/Resources/src/ThermalZones/EnergyPlus/C-Sources/InputVariableAllocate.h
@@ -25,7 +25,7 @@ LBNL_EnergyPlus_EXPORT void* EnergyPlusInputVariableAllocate(
   const char* fmuName,
   const char* buildingsLibraryRoot,
   const double initialTime,
-  const int logLevel,
+  const size_t logLevel,
   void (*SpawnMessage)(const char *string),
   void (*SpawnError)(const char *string),
   void (*SpawnFormatMessage)(const char *string, ...),
